+++ public/front-roms-library/src/components/VideoconsoleFilter/VideoconsoleFilter.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { OnSwitchVideoconsolesSelected, Videoconsole } from "../../types"
 
 interface VideoconsoleFilterProps {
@@ -6,25 +6,25 @@ interface VideoconsoleFilterProps {
     onSwitchVideoconsolesSelected: OnSwitchVideoconsolesSelected
 }
 
-export const VideoconsoleFilter: React.FC<VideoconsoleFilterProps> = ({videoconsole, onSwitchVideoconsolesSelected}) => {
+export const VideoconsoleFilter: React.FC<VideoconsoleFilterProps> = ({ videoconsole, onSwitchVideoconsolesSelected}) => {
 
     const [inputChecked, setInputChecked] = useState<boolean>(true);
 
     const onInputChange = () => {
         setInputChecked(!inputChecked);
-        onSwitchVideoconsolesSelected(videoconsole, inputChecked);
+        onSwitchVideoconsolesSelected(videoconsole, !inputChecked);
     }
 
     return (
         <>
             <label htmlFor={videoconsole.slug}>{videoconsole.name}</label>
-            <input 
-                type="checkbox" 
-                name={videoconsole.name} 
-                id={videoconsole.slug}  
+            <input
+                type="checkbox"
+                name={videoconsole.name}
+                id={videoconsole.slug}
                 onChange={onInputChange}
                 checked={inputChecked}
-                />
+            />
         </>
     )
 }