+++ public/front-roms-library/src/helpers/getConsoles.ts
@@ -1,22 +1,27 @@
 import { Videoconsole, PostConsole } from "../types";
 import { getApiHost } from "./";
 
-export const getConsoles = async() => {
+export const getConsoles = async () => {
     // TODO - In the future could be more than 10 consoles. Control pagination.
-    const url = getApiHost() + '/wp-json/wp/v2/console/?acf_format=standard'
-    const resp = await fetch(url);
-    const data = await resp.json();
-    const consoles: Videoconsole[] = data.map((console: PostConsole) => {
-        return {
-            id: console.id,
-            name: console.name,
-            description: console.description,
-            slug: console.slug,
-            image: false !== console.acf.console_image ? console.acf.console_image : getImageBySlug(console.slug),
-        }
-    })
-    
-    return consoles;
+    const url = getApiHost() + '/wp-json/wp/v2/console/?acf_format=standard';
+    return fetch(url)
+        .then(resp => {
+            if (!resp.ok) {
+                throw new Error(resp.statusText);
+            }
+            return resp.json() as Promise<PostConsole[]>;
+        }).then(data => {
+            const consoles: Videoconsole[] = data.map((console: PostConsole) => {
+                return {
+                    id: console.id,
+                    name: console.name,
+                    description: console.description,
+                    slug: console.slug,
+                    image: false !== console.acf.console_image ? console.acf.console_image : getImageBySlug(console.slug),
+                }
+            })
+            return consoles as Videoconsole[];
+        })
 }
 
 const getImageBySlug = (slug: string) => {