+++ public/front-roms-library/src/types.tsx
@@ -1,12 +1,12 @@
 export type OnNewRom = (newRom: Rom) => void;
 export type GetNewRoms = (getRomsArgs: getRomsArgs) => void;
-export type OnSetVideoconsole = (videoconsole: string) => void;
+export type OnSetVideoconsole = (videoconsole: Videoconsole[]) => void;
 export type OnSwitchVideoconsolesSelected = (videoconsole: Videoconsole, isConsoleFiltered: boolean) => void;
 
 export interface getRomsArgs {
     term?: string,
     page?: number, 
-    videoconsole?: string
+    videoconsoles?: Videoconsole[]
 }
 export interface PostRom {
     acf: {